+++ src/fhirpath.js
@@ -96,8 +96,9 @@ engine.invocationTable = {
   combine:      {fn: combining.combineFn, arity: {1: ["AnyAtRoot"]}},
   union:        {fn: combining.union,   arity: {1: ["AnyAtRoot"]}},
   intersect:    {fn: combining.intersect,   arity: {1: ["AnyAtRoot"]}},
+  exclude:      {fn: combining.exclude,   arity: {1: ["AnyAtRoot"]}},
   iif:          {fn: misc.iifMacro,    arity: {2: ["Expr", "Expr"], 3: ["Expr", "Expr", "Expr"]}},
-  trace:        {fn: misc.traceFn,     arity: {0: [], 1: ["String"]}},
+  trace:        {fn: misc.traceFn,     arity: {1: ["String"], 2: ["String", "Expr"]}},
   toInteger:    {fn: misc.toInteger},
   toDecimal:    {fn: misc.toDecimal},
   toString:     {fn: misc.toString},
@@ -106,6 +107,7 @@ engine.invocationTable = {
   toTime:       {fn: misc.toTime},
   toBoolean:    {fn: misc.toBoolean},
   toQuantity:   {fn: misc.toQuantity, arity: {0: [], 1: ["String"]}},
+  hasValue:     {fn: misc.hasValueFn},
   convertsToBoolean:    {fn: misc.createConvertsToFn(misc.toBoolean, 'boolean')},
   convertsToInteger:    {fn: misc.createConvertsToFn(misc.toInteger, 'number')},
   convertsToDecimal:    {fn: misc.createConvertsToFn(misc.toDecimal, 'number')},
@@ -127,6 +129,12 @@ engine.invocationTable = {
   replaceMatches: {fn: strings.replaceMatches,   arity: {2: ["String", "String"]}},
   length:         {fn: strings.length },
   toChars:        {fn: strings.toChars },
+  join:           {fn: strings.joinFn,           arity: {0: [], 1: ["String"]}},
+  split:          {fn: strings.splitFn,          arity: {1: ["String"]}},
+  trim:           {fn: strings.trimFn},
+
+  encode:         {fn: strings.encodeFn,         arity: {1: ["String"]}},
+  decode:         {fn: strings.decodeFn,         arity: {1: ["String"]}},
 
   abs:            {fn: math.abs},
   ceiling:        {fn: math.ceiling},
@@ -372,6 +380,18 @@ engine.MemberInvocation = function(ctx, parentData, node ) {
           if(Array.isArray(toAdd)) {
             acc = acc.concat(toAdd.map((x, i)=>
               makeResNode(x, childPath, _toAdd && _toAdd[i])));
+            // Add items to the end of the ResourceNode list that have no value
+            // but have associated data, such as extensions or ids.
+            const _toAddLength = _toAdd?.length || 0;
+            for (let i = toAdd.length; i < _toAddLength; ++i) {
+              acc.push(makeResNode(null, childPath, _toAdd[i]));
+            }
+          } else if (toAdd == null && Array.isArray(_toAdd)) {
+            // Add items to the end of the ResourceNode list when there are no
+            // values at all, but there is a list of associated data, such as
+            // extensions or ids.
+            acc = acc.concat(_toAdd.map((x) =>
+              makeResNode(null, childPath, x)));
           } else {
             acc.push(makeResNode(toAdd, childPath, _toAdd));
           }
@@ -423,13 +443,13 @@ engine.realizeParams = function(ctx, parentData, args) {
 function makeParam(ctx, parentData, type, param) {
   if(type === "Expr"){
     return function(data) {
-      ctx.$this = util.arraify(data);
-      return engine.doEval(ctx, ctx.$this, param);
+      const $this = util.arraify(data);
+      return engine.doEval({ ...ctx, $this }, $this, param);
     };
   }
   if(type === "AnyAtRoot"){
-    ctx.$this = ctx.$this || ctx.dataRoot;
-    return engine.doEval(ctx, ctx.$this, param);
+    const $this = ctx.$this || ctx.dataRoot;
+    return engine.doEval({ ...ctx, $this}, $this, param);
   }
   if(type === "Identifier"){
     if(param.type === "TermExpression") {
@@ -443,7 +463,6 @@ function makeParam(ctx, parentData, type, param) {
     return engine.TypeSpecifier(ctx, parentData, param);
   }
 
-  ctx.$this = parentData;
   const res = engine.doEval(ctx, parentData, param);
   if(type === "Any") {
     return res;
@@ -459,7 +478,7 @@ function makeParam(ctx, parentData, type, param) {
 }
 
 function doInvoke(ctx, fnName, data, rawParams){
-  var invoc = engine.invocationTable[fnName];
+  var invoc = ctx.userInvocationTable?.[fnName] ?? engine.invocationTable[fnName];
   var res;
   if(invoc) {
     if(!invoc.arity){
@@ -640,9 +659,12 @@ function parse(path) {
  * @param {object} context - a hash of variable name/value pairs.
  * @param {object} model - The "model" data object specific to a domain, e.g. R4.
  *  For example, you could pass in the result of require("fhirpath/fhir-context/r4");
- * @param {object} [options] - additional options:
+ * @param {object} options - additional options:
  * @param {boolean} [options.resolveInternalTypes] - whether values of internal
  *  types should be converted to strings, true by default.
+ * @param {function} [options.traceFn] - An optional trace function to call when tracing.
+ * @param {object} [options.userInvocationTable] - a user invocation table used
+ *  to replace any existing or define new functions.
  */
 function applyParsedPath(resource, parsedPath, context, model, options) {
   constants.reset();
@@ -666,13 +688,19 @@ function applyParsedPath(resource, parsedPath, context, model, options) {
     }, {});
   }
   let ctx = {dataRoot, vars: Object.assign(vars, context), model};
+  if (options.traceFn) {
+    ctx.customTraceFn = options.traceFn;
+  }
+  if (options.userInvocationTable) {
+    ctx.userInvocationTable = options.userInvocationTable;
+  }
   return  engine.doEval(ctx, dataRoot, parsedPath.children[0])
     // engine.doEval returns array of "ResourceNode" and/or "FP_Type" instances.
     // "ResourceNode" or "FP_Type" instances are not created for sub-items.
     // Resolve any internal "ResourceNode" instances to plain objects and if
     // options.resolveInternalTypes is true, resolve any internal "FP_Type"
     // instances to strings.
-    .map(n => {
+    .reduce((acc,n) => {
       // Path for the data extracted from the resource.
       let path = n instanceof ResourceNode ? n.path : null;
       n = util.valData(n);
@@ -681,13 +709,17 @@ function applyParsedPath(resource, parsedPath, context, model, options) {
           n = n.toString();
         }
       }
-      // Add a hidden (non-enumerable) property with the path to the data extracted
-      // from the resource.
-      if (path && typeof n === 'object') {
-        Object.defineProperty(n, '__path__', {value: path});
+      // Exclude nulls
+      if (n != null) {
+        // Add a hidden (non-enumerable) property with the path to the data extracted
+        // from the resource.
+        if (path && typeof n === 'object') {
+          Object.defineProperty(n, '__path__', {value: path});
+        }
+        acc.push(n);
       }
-      return n;
-    });
+      return acc;
+    }, []);
 }
 
 /**
@@ -721,14 +753,17 @@ function resolveInternalTypes(val) {
  *  or object, if fhirData represents the part of the FHIR resource:
  * @param {string} path.base - base path in resource from which fhirData was extracted
  * @param {string} path.expression - FHIRPath expression relative to path.base
- * @param {object} context - a hash of variable name/value pairs.
- * @param {object} model - The "model" data object specific to a domain, e.g. R4.
+ * @param {object} [context] - a hash of variable name/value pairs.
+ * @param {object} [model] - The "model" data object specific to a domain, e.g. R4.
  *  For example, you could pass in the result of require("fhirpath/fhir-context/r4");
  * @param {object} [options] - additional options:
  * @param {boolean} [options.resolveInternalTypes] - whether values of internal
  *  types should be converted to standard JavaScript types (true by default).
  *  If false is passed, this conversion can be done later by calling
  *  resolveInternalTypes().
+ * @param {function} [options.traceFn] - An optional trace function to call when tracing.
+ * @param {object} [options.userInvocationTable] - a user invocation table used
+ *  to replace any existing or define new functions.
  */
 function evaluate(fhirData, path, context, model, options) {
   return compile(path, model, options)(fhirData, context);
@@ -744,17 +779,45 @@ function evaluate(fhirData, path, context, model, options) {
  * @param {string} path.base - base path in resource from which a part of
  *   the resource was extracted
  * @param {string} path.expression - FHIRPath expression relative to path.base
- * @param {object} model - The "model" data object specific to a domain, e.g. R4.
+ * @param {object} [model] - The "model" data object specific to a domain, e.g. R4.
  *  For example, you could pass in the result of require("fhirpath/fhir-context/r4");
  * @param {object} [options] - additional options:
  * @param {boolean} [options.resolveInternalTypes] - whether values of internal
  *  types should be converted to strings, true by default.
+ * @param {function} [options.traceFn] - An optional trace function to call when tracing.
+ * @param {object} [options.userInvocationTable] - a user invocation table used
+ *  to replace any existing or define new functions.
  */
 function compile(path, model, options) {
   options = {
     resolveInternalTypes: true,
     ... options
   };
+
+  const userInvocationTable = options.userInvocationTable;
+  if (userInvocationTable) {
+    options.userInvocationTable = Object.keys(userInvocationTable).reduce(
+      (invocationTable, fnName) => {
+        if (userInvocationTable[fnName].internalStructures) {
+          invocationTable[fnName] = userInvocationTable[fnName];
+        } else {
+          invocationTable[fnName] = {
+            ...userInvocationTable[fnName],
+            fn: (...args) => {
+              return userInvocationTable[fnName].fn.apply(
+                // When we check Array.isArray(arg), we are checking if the
+                // singleton function has been called. An alternative to this is
+                // to check that the type of the argument is Integer, Boolean,
+                // Number, or String.
+                this, args.map(arg => Array.isArray(arg) ? arg.map(item => util.valData(item)) : arg)
+              );
+            }
+          };
+        }
+        return invocationTable;
+      }, {});
+  }
+
   if (typeof path === 'object') {
     const node = parse(path.expression);
     return function (fhirData, context) {
@@ -794,5 +857,7 @@ module.exports = {
   resolveInternalTypes,
   types: typesFn,
   // Might as well export the UCUM library, since we are using it.
-  ucumUtils: require('@lhncbc/ucum-lhc').UcumLhcUtils.getInstance()
+  ucumUtils: require('@lhncbc/ucum-lhc').UcumLhcUtils.getInstance(),
+  // Utility functions that can be used to implement custom functions
+  util
 };