+++ test/benchmark.js
@@ -20,7 +20,8 @@ const open = require('open');
 const current_fhirpath = require('../src/fhirpath');
 const current_r4_model = require('../fhir-context/r4');
 const minimumDataset = require('./resources/Minimum-Data-Set---version-3.0.R4.json');
-const currentVersion = require('../package.json').version;
+const _ = require('lodash');
+const currentVersion = 'current';
 
 
 const child = spawn('npm', ['i','--prefix', './test/benchmark/prev-fhirpath', 'fhirpath' + prevVersion], {
@@ -32,12 +33,21 @@ child.on('exit', code => {
     const previous_fhirpath = require('./benchmark/prev-fhirpath/node_modules/fhirpath');
     const previous_r4_model = require('./benchmark/prev-fhirpath/node_modules/fhirpath/fhir-context/r4');
     const previousVersion = require('./benchmark/prev-fhirpath/node_modules/fhirpath/package.json').version;
+    const bigItems = current_fhirpath.evaluate(minimumDataset,'repeat(item)', {},  current_r4_model);
+    const bigItemsCopy = _.cloneDeep(bigItems);
+    const smallItems = current_fhirpath.evaluate(minimumDataset,'repeat(item).repeat(code)', {},  current_r4_model)
+    const smallItemsCopy = _.cloneDeep(smallItems);
     // Insert performance test suites here:
     [
       'intersect',
-      'member-invocation'
+      'member-invocation',
+      'union',
+      'exclude',
+      'subsetof',
+      'distinct',
+      'contains'
     ].forEach(filename => {
-      require('./benchmark/'+ filename)({
+      const suites = require('./benchmark/'+ filename)({
         benny,
         open,
         previous_fhirpath,
@@ -46,8 +56,53 @@ child.on('exit', code => {
         current_r4_model,
         minimumDataset,
         currentVersion,
-        previousVersion
-      })
+        previousVersion,
+        bigItems,
+        bigItemsCopy,
+        smallItems,
+        smallItemsCopy
+      });
+      suites.forEach(suite => {
+        const cases = suite.cases.reduce((arr, item) => {
+          arr.push(
+            benny.add(
+              `${item.name} [${previousVersion}]`,
+              item.testFunction.bind(
+                this,
+                previous_fhirpath,
+                previous_r4_model,
+                previous_fhirpath.compile(suite.expression, previous_r4_model)
+              )
+            )
+          );
+          arr.push(
+            benny.add(
+              `${item.name} [${currentVersion}]`,
+              item.testFunction.bind(
+                this,
+                current_fhirpath,
+                current_r4_model,
+                current_fhirpath.compile(suite.expression, current_r4_model)
+              )
+            )
+          );
+          return arr;
+        }, []);
+
+        benny.suite(
+          suite.name,
+          ...cases,
+          benny.cycle(),
+          benny.complete(),
+          benny.configure({
+            minDisplayPrecision: 2
+          }),
+          benny.save({ file: suite.filename, folder: __dirname + '/results', version: currentVersion }),
+          benny.save({ file: suite.filename, folder: __dirname + '/results', format: 'chart.html' }),
+        ).then(() => {
+          open(__dirname + `/results/${suite.filename}.chart.html`);
+        });
+      });
     })
   }
 });